Encapsulate home page data filters
